@@ -11,26 +11,34 @@
     "type-check": "tsc --noEmit --isolatedDeclarations",
     "new-post": "node scripts/new-post.js",
     "format": "biome format --write ./src",
-    "lint": "biome check --apply ./src"
+    "lint": "biome check --apply ./src",
+    "preinstall": "npx only-allow pnpm"
   },
   "dependencies": {
-    "@astrojs/check": "^0.9.3",
-    "@astrojs/rss": "^4.0.7",
-    "@astrojs/sitemap": "^3.1.6",
-    "@astrojs/svelte": "^5.7.1",
-    "@astrojs/tailwind": "^5.1.1",
-    "@fontsource-variable/jetbrains-mono": "^5.0.22",
-    "@fontsource/roboto": "^5.0.14",
+    "@astrojs/check": "^0.9.4",
+    "@astrojs/rss": "^4.0.9",
+    "@astrojs/sitemap": "^3.2.1",
+    "@astrojs/svelte": "^6.0.1",
+    "@astrojs/tailwind": "^5.1.2",
+    "@fontsource-variable/jetbrains-mono": "^5.1.1",
+    "@fontsource/roboto": "^5.1.0",
+    "@iconify-json/fa6-brands": "^1.2.1",
+    "@iconify-json/fa6-regular": "^1.2.1",
+    "@iconify-json/fa6-solid": "^1.2.1",
+    "@iconify-json/material-symbols": "^1.2.5",
+    "@iconify/svelte": "^4.0.2",
+    "@sveltejs/vite-plugin-svelte": "^4.0.0",
     "@swup/astro": "^1.4.1",
-    "astro": "^4.15.9",
-    "astro-compress": "^2.3.1",
+    "@tailwindcss/typography": "^0.5.15",
+    "astro": "^4.16.13",
+    "astro-compress": "^2.3.5",
     "astro-icon": "^1.1.1",
-    "colorjs.io": "^0.5.2",
     "hastscript": "^9.0.0",
     "markdown-it": "^14.1.0",
     "mdast-util-to-string": "^4.0.0",
     "overlayscrollbars": "^2.10.0",
-    "pagefind": "^1.1.0",
+    "pagefind": "^1.1.1",
+    "photoswipe": "^5.4.4",
     "reading-time": "^1.5.0",
     "rehype-autolink-headings": "^7.1.0",
     "rehype-components": "^0.3.0",
@@ -38,37 +46,25 @@
     "rehype-slug": "^6.0.0",
     "remark-directive": "^3.0.0",
     "remark-directive-rehype": "^0.4.2",
+    "remark-github-admonitions-to-directives": "^1.0.5",
     "remark-math": "^6.0.0",
-    "sanitize-html": "^2.13.0",
+    "sanitize-html": "^2.13.1",
     "sharp": "^0.33.5",
-    "svelte": "^4.2.19",
-    "tailwindcss": "^3.4.10",
-    "typescript": "^5.5.4",
+    "stylus": "^0.63.0",
+    "svelte": "^5.2.2",
+    "tailwindcss": "^3.4.14",
+    "typescript": "^5.6.3",
     "unist-util-visit": "^5.0.0"
   },
   "devDependencies": {
-    "@astrojs/ts-plugin": "^1.10.2",
+    "@astrojs/ts-plugin": "^1.10.4",
     "@biomejs/biome": "1.8.3",
-    "@iconify-json/fa6-brands": "^1.1.22",
-    "@iconify-json/fa6-regular": "^1.1.22",
-    "@iconify-json/fa6-solid": "^1.1.24",
-    "@iconify-json/material-symbols": "^1.1.89",
-    "@iconify/svelte": "^4.0.2",
     "@rollup/plugin-yaml": "^4.1.2",
-    "@tailwindcss/typography": "^0.5.15",
     "@types/markdown-it": "^14.1.2",
     "@types/mdast": "^4.0.4",
     "@types/sanitize-html": "^2.13.0",
-    "photoswipe": "^5.4.4",
-    "remark-github-admonitions-to-directives": "^1.0.5",
-    "sass": "^1.77.8",
-    "stylus": "^0.63.0"
+    "postcss-import": "^16.1.0",
+    "postcss-nesting": "^13.0.1"
   },
-  "pnpm": {
-    "overrides": {
-      "vite-imagetools": "^6.2.7",
-      "sharp": "^0.33.0"
-    }
-  },
-  "packageManager": "pnpm@9.6.0+sha512.38dc6fba8dba35b39340b9700112c2fe1e12f10b17134715a4aa98ccf7bb035e76fd981cf0bb384dfa98f8d6af5481c2bef2f4266a24bfa20c34eb7147ce0b5e"
+  "packageManager": "pnpm@9.14.4"
 }
